Organize main nav files
