+++ src/components/components.js
@@ -28,9 +28,9 @@ import LinkGroup from './navigation/linkGroup';
 import LinkList from './simple/linkList';
 import LocalFooter from './identity/localFooter';
 import LocalFooterPicker from './identity/localFooterPicker';
-import MainMenuGroup from './navigation/mainMenuGroup';
-import MainNav from './navigation/mainNav';
-import MainNavItem from './navigation/mainNavItem';
+import MainMenuGroup from './navigation/MainNav/mainMenuGroup';
+import MainNav from './navigation/MainNav/mainNav';
+import MainNavItem from './navigation/MainNav/mainNavItem';
 import Masthead from './identity/masthead';
 import MastheadPicker from './identity/mastheadPicker';
 import NavItem from './navigation/navItem';
