+++ src/components/storyblok/saaMainNav.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import SbEditable from 'storyblok-react';
-import SAAMainNav from '../navigation/SAAMainNav/SAAMainNav';
+import SAAMainNav from '../navigation/MainNav/SAAMainNav';
 
 export const SBSAAMainNav = ({ blok, ariaLabel }) => {
   const { menuItems } = blok;
